CTAA-1660 - Remove code related to download and managing of the received infections
