+++ app/src/main/java/at/roteskreuz/stopcorona/di/PersistenceModule.kt
@@ -36,7 +36,6 @@ internal val persistenceModule = module {
         DatabaseCleanupManagerImpl(
             appDispatchers = get(),
             configurationRepository = get(),
-            infectionMessageDao = get(),
             temporaryExposureKeysDao = get()
         )
     }
@@ -55,10 +54,6 @@ internal val persistenceModule = module {
 
     single {
         get<DefaultDatabase>().sessionDao()
-    }
-
-    single {
-        get<DefaultDatabase>().infectionMessageDao()
     }
 
     single {
