+++ app/src/main/java/at/roteskreuz/stopcorona/di/ViewModelModule.kt
@@ -34,8 +34,7 @@ val viewModelModule = module {
             appDispatchers = get(),
             infectionMessengerRepository = get(),
             notificationsRepository = get(),
-            quarantineRepository = get(),
-            infectionMessageDao = get()
+            quarantineRepository = get()
         )
     }
