+++ app/src/main/java/at/roteskreuz/stopcorona/model/api/ApiDescription.kt
@@ -2,8 +2,9 @@ package at.roteskreuz.stopcorona.model.api
 
 import at.roteskreuz.stopcorona.model.entities.configuration.ApiConfigurationHolder
 import at.roteskreuz.stopcorona.model.entities.infection.info.ApiInfectionDataRequest
-import at.roteskreuz.stopcorona.model.entities.infection.message.ApiInfectionMessages
-import retrofit2.http.*
+import retrofit2.http.Body
+import retrofit2.http.GET
+import retrofit2.http.POST
 
 /**
  * Description of REST Api for Retrofit.
@@ -12,10 +13,6 @@ interface ApiDescription {
 
     @GET("configuration")
     suspend fun configuration(): ApiConfigurationHolder
-
-    @Deprecated(message = "This is API is obsolete, we will use the Exposure Notificationframework")
-    @GET("infection-messages")
-    suspend fun infectionMessages(@Query("addressPrefix") addressPrefix: String, @Query("fromId") fromId: Long? = null): ApiInfectionMessages
 
     @POST("publish")
     suspend fun publish(@Body infectionDataRequest: ApiInfectionDataRequest)
