+++ app/src/main/java/at/roteskreuz/stopcorona/model/entities/exposure/DbSentTemporaryExposureKeys.kt
@@ -3,10 +3,11 @@ package at.roteskreuz.stopcorona.model.entities.exposure
 import android.os.Parcelable
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import androidx.room.TypeConverter
 import at.roteskreuz.stopcorona.model.entities.infection.message.MessageType
 import at.roteskreuz.stopcorona.skeleton.core.model.entities.DbEntity
 import kotlinx.android.parcel.Parcelize
-import java.util.*
+import java.util.UUID
 
 /**
  * Describes a sent temporary exposure key through it's associated rolling start interval number
@@ -39,5 +40,34 @@ data class DbSentTemporaryExposureKeys(
         result = 31 * result + rollingStartIntervalNumber.hashCode()
         result = 31 * result + messageType.hashCode()
         return result
+    }
+}
+
+/**
+ * Converter from and to UUID.
+ */
+class UUIDConverter {
+
+    @TypeConverter
+    fun stringToUUID(value: String?): UUID? =
+        value?.let { UUID.fromString(it) }
+
+    @TypeConverter
+    fun uuidToString(uuid: UUID?): String? = uuid?.toString()
+}
+
+/**
+ * Converter from and to [MessageType].
+ */
+class MessageTypeConverter {
+
+    @TypeConverter
+    fun toIdentifier(messageType: MessageType?): String? {
+        return messageType?.identifier
+    }
+
+    @TypeConverter
+    fun fromIdentifier(identifier: String?): MessageType? {
+        return identifier?.let { MessageType(it) }
     }
 }
